updated test according to changes
